chore (users | front): moved `UserForm` to `components/`
